// src/app/app-resources/spine/match-type.ts
export enum MATCH_TYPE{
    SWISS = 0,
    SWISS_ROUND_1 = 1,
    SWISS_ROUND_2 = 2,
    SWISS_ROUND_3 = 3,
    SWISS_ROUND_4 = 4,
    SWISS_ROUND_5 = 5,
    SWISS_ROUND_6 = 6,
    SWISS_ROUND_7 = 7,
    SWISS_ROUND_8 = 8,
    TOP_132 = 9,
    TOP_64 = 10,
    TOP_32 = 11,
    TOP_16 = 12,
    QUATER_FINALS = 13,
    SEMI_FINALS = 14,
    FINALS = 15
}